# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Minimarket Plus API",
        "description": "API REST del backend de Minimarket Plus: productos, carrito, inventario, ventas y usuarios. Documentada con OpenAPI e hipermedios HATEOAS. Los recursos est\u00e1n protegidos por roles (CLIENTE, CAJERO, ADMIN).",
        "contact": {
            "name": "Equipo Minimarket Plus"
        },
        "license": {
            "name": "Uso acad\u00e9mico - Duoc UC"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8084",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/inventario/{id}": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Obtener movimiento por ID",
                "operationId": "obtenerMovimientoPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Movimiento no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelInventario"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Movimiento encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelInventario"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Actualizar movimiento de inventario",
                "operationId": "actualizarMovimiento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventario"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Movimiento no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelInventario"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Movimiento actualizado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelInventario"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Eliminar movimiento de inventario",
                "operationId": "eliminarMovimiento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Movimiento no encontrado"
                    },
                    "204": {
                        "description": "Movimiento eliminado"
                    }
                }
            }
        },
...
        "/api/inventario": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Listar movimientos de inventario",
                "description": "Devuelve todos los movimientos de stock con enlaces HATEOAS.",
                "operationId": "listarMovimientosDeInventario",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionModelEntityModelInventario"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos (ADMIN o CAJERO)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionModelEntityModelInventario"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Registrar movimiento de inventario",
                "description": "Registra una entrada o salida de stock asociada a un producto.",
                "operationId": "registrarMovimiento",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventario"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Movimiento registrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelInventario"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos (ADMIN o CAJERO)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelInventario"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/ventas/{id}": {
            "get": {
                "tags": [
                    "venta-controller"
                ],
                "operationId": "obtenerVentaPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Venta"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Rol": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Usuario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "roles": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rol"
                        }
                    }
                }
            },
...
            "Links": {
                "type": "object",
                "additionalProperties": {
                    "$ref": "#/components/schemas/Link"
                }
            },
            "Categoria": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Producto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                }
            },
...
            "Inventario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "cantidad": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "tipoMovimiento": {
                        "type": "string"
                    },
                    "fechaMovimiento": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "EntityModelInventario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "cantidad": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "tipoMovimiento": {
                        "type": "string"
                    },
                    "fechaMovimiento": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
            "DetalleVenta": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "cantidad": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "Venta": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "detalles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetalleVenta"
                        }
                    }
                }
            },
...
            "CollectionModelEntityModelInventario": {
                "type": "object",
                "properties": {
                    "_embedded": {
                        "type": "object",
                        "properties": {
                            "inventarioList": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/EntityModelInventario"
                                }
                            }
                        }
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
...
            "Link": {
                "type": "object",
                "properties": {
                    "href": {
                        "type": "string"
                    },
                    "hreflang": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "deprecation": {
                        "type": "string"
                    },
                    "profile": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "templated": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}